requirements.txt
streamlit
fastapi
uvicorn
pandas
numpy
python-dotenv
pydantic
scikit-learn
tensorflow
plotly